requirements.txt
fastapi
uvicorn
jinja2
psycopg2-binary
watchdog
joblib
ray
tensorflow
pyyaml